Don't use short option to get last_commit
Short commit hashes are difficult to compare and if we end up
with hashes of different length we can compare them using
start_with?, but since there can be commits with similar prefixes
we can be wrong comparing full-length hash and some short-length
hash with same prefixes.

Example:

```
$ ls $(brew --cellar)/ack
HEAD-b090608c
```

Let's now suppose we wan't to check if HEAD-b09060c is outdated
using GitHub API, which return full-length prefix and let's
suppose it returns the following hash:

b090608c49643099a47247d79ac9ca80d05c2ee0

How do we treat it? Is the HEAD version outdated or not? Suppose
it's not since prefixes are the same. After a while we check it
again and GitHub API returns the following hash:

b090608c00000000000000000000000000000000

So, we still think our installed HEAD is up-to-date, however
that's not true and our HEAD is old as the hills.