cmd/install: fix checks for installed formulae
`f.installed?` means that last prefix is installed for formula spec.
However it seems to me that here it's treated as a check for any
version being installed.

`current ||= f.old_installed_formulae.first` doesn't always mean that
`f.old_installed_formulae.first.installed?` is true.

The fix is to check if `f.installed?` is true or old formulae installed
with `f` alias is `installed?` which implies that it's latest prefix
is installed.